requirements.txt
requests==2.31.0
python-ulid==2.2.0
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.2.1
xlrd==2.0.1
openpyxl==3.1.2
sqlalchemy==2.0.30
tenacity==8.3.0
psutil==6.1.0
duckdb==1.1.3

pre-commit==3.6.2
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
